backend/requirements.txt
fastapi
uvicorn[standard]
gunicorn
python-dotenv
feedparser
httpx
pydantic-settings
sqlalchemy[asyncio]
aiosqlite
apscheduler
google-generativeai
